switched to dynamic playstore detection
